refactor(api): move tooling integration tests into tooling
